archive/tts-factory/server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-cloud-texttospeech>=2.16.0
pydantic>=2.10.0
python-multipart>=0.0.6
